requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asn1crypto==0.24.0        # via cryptography
bcrypt==3.1.7             # via paramiko
cffi==1.12.3              # via bcrypt, cryptography, pynacl
click==7.0                # via pip-tools
cryptography==2.7         # via paramiko
django-storages==1.7.1
django==2.2.4
fabric3==1.14.post1
paramiko==2.6.0           # via fabric3
pip-tools==4.0.0
psycopg2-binary==2.8.3
pycparser==2.19           # via cffi
pynacl==1.3.0             # via paramiko
pytz==2019.2              # via django
six==1.12.0               # via bcrypt, cryptography, fabric3, pip-tools, pynacl
sqlparse==0.3.0           # via django
